Make guide items collapsible
Make it easier to see many items by making them collapsible.
