+++ src/data-guide/DataGuide.tsx
@@ -1,17 +1,7 @@
-import { createContext, useContext, useEffect, useMemo, useState } from "react";
 import { cn } from "@/lib/utils";
+import { useCallback, useState } from "react";
 import { Sidebar } from "./Sidebar";
-import { HelpItemId, useTrackGuideItems } from "./items";
-
-interface DataGuideContext {
-  pushGuideItems: (items: HelpItemId[]) => void;
-  popGuideItems: (items: HelpItemId[]) => void;
-}
-
-const DataGuideContext = createContext<DataGuideContext>({
-  pushGuideItems: () => {},
-  popGuideItems: () => {},
-});
+import { DataGuideContext, useTrackGuideItems } from "./internal";
 
 export const DataGuideProvider = ({
   children,
@@ -23,9 +13,9 @@ export const DataGuideProvider = ({
 
   const showSidebar = items.length > 0;
 
-  const toggleOpen = () => {
-    setOpen(!open);
-  };
+  const toggleOpen = useCallback(() => {
+    setOpen((isOpen) => !isOpen);
+  }, [setOpen]);
 
   return (
     <DataGuideContext.Provider
@@ -44,29 +34,4 @@ export const DataGuideProvider = ({
       )}
     </DataGuideContext.Provider>
   );
-};
-
-/**
- * The main way for components to indicate which help items are relevant when the component is used.
- *
- * @param items The help items relevant to the component using this hook
- */
-export const useDataGuide = (items: HelpItemId[]) => {
-  const context = useContext(DataGuideContext);
-  if (!context) {
-    throw new Error("useDataGuide must be used within a DataGuideProvider");
-  }
-
-  const { pushGuideItems, popGuideItems } = context;
-
-  // Only use the initial values and ignore any changes to it
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  const setupItems = useMemo(() => items, []);
-
-  useEffect(() => {
-    pushGuideItems(setupItems);
-    return () => {
-      popGuideItems(setupItems);
-    };
-  }, [setupItems, pushGuideItems, popGuideItems]);
 };
